+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/LogicalTable.java
@@ -7,18 +7,26 @@ import com.yahoo.bard.webservice.data.metric.LogicalMetric;
 import com.yahoo.bard.webservice.data.metric.LogicalMetricColumn;
 import com.yahoo.bard.webservice.data.metric.MetricDictionary;
 import com.yahoo.bard.webservice.data.time.Granularity;
+import com.yahoo.bard.webservice.web.filters.ApiFilters;
 
 import org.joda.time.ReadablePeriod;
 import org.joda.time.Years;
 
 import java.util.LinkedHashSet;
+import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
 import javax.validation.constraints.NotNull;
 
 /**
- * A LogicalTable has a grain and a tablegroup of physical tables that satisfy the logical table.
+ * A LogicalTable can be conceived as a view on the underlying physical dataSources that it supports.
+ *
+ * The LogicalTableSchema defines the grains and columns available on that view.  The viewFilters field provides row
+ * filtering on that view.
+ *
+ * The TableGroup field provides a collection of tables that should, collectively, satisfy the schema of this
+ * LogicalTable.  During query planning that collection is scanned and compared for lowest row cardinality.
  */
 public class LogicalTable implements Table, Comparable<LogicalTable> {
 
@@ -26,13 +34,15 @@ public class LogicalTable implements Table, Comparable<LogicalTable> {
     public static final ReadablePeriod DEFAULT_RETENTION = Years.ONE;
 
     private String name;
-    private TableGroup tableGroup;
-    private LogicalTableSchema schema;
 
     private String category;
     private String longName;
     private ReadablePeriod retention;
     private String description;
+
+    private TableGroup tableGroup;
+    private LogicalTableSchema schema;
+    private ApiFilters viewFilters;
 
     // parameter used by the compare to method
     private String comparableParam;
@@ -42,24 +52,31 @@ public class LogicalTable implements Table, Comparable<LogicalTable> {
      * <p>
      * Sets the Category and Retention to the defaults, and sets the long name and description to the name.
      *
+     * Defaults to the TableGroup for schema and no ApiFilters.
+     *
      * @param name  The logical table name
      * @param granularity  The logical table granularity
      * @param tableGroup  The tablegroup for the logical table
      * @param metricDictionary The metric dictionary to bind tableGroup's metrics
+     *
+     * @deprecated prefer constructor that uses the LogicalTableName class
      */
+    @Deprecated
     public LogicalTable(
             @NotNull String name,
             @NotNull Granularity granularity,
             TableGroup tableGroup,
             MetricDictionary metricDictionary
     ) {
-        this(name, DEFAULT_CATEGORY, name, granularity, DEFAULT_RETENTION, name, tableGroup, metricDictionary);
+        this(LogicalTableName.forName(name), granularity, tableGroup, metricDictionary);
     }
 
     /**
      * Constructor.
      *
      * Uses the LogicalTableName interface to package metadata.
+     *
+     * Defaults to TableGroup for schema and no ApiFilters.
      *
      * @param name  The logical table name
      * @param granularity  The logical table time grain
@@ -87,6 +104,8 @@ public class LogicalTable implements Table, Comparable<LogicalTable> {
     /**
      * Constructor.
      *
+     * Defaults to the TableGroup schema.
+     *
      * @param name  The logical table name
      * @param category  The category of the logical table
      * @param longName  The long name of the logical table
@@ -95,6 +114,7 @@ public class LogicalTable implements Table, Comparable<LogicalTable> {
      * @param description  The description for this logical table
      * @param tableGroup  The tablegroup for the logical table
      * @param metricDictionary The metric dictionary to bind tableGroup's metrics
+     *
      */
     public LogicalTable(
             @NotNull String name,
@@ -114,7 +134,8 @@ public class LogicalTable implements Table, Comparable<LogicalTable> {
                 retention,
                 description,
                 tableGroup,
-                new LogicalTableSchema(tableGroup, granularity, metricDictionary)
+                new LogicalTableSchema(tableGroup, granularity, metricDictionary),
+                tableGroup.getApiFilters()
         );
     }
 
@@ -129,6 +150,7 @@ public class LogicalTable implements Table, Comparable<LogicalTable> {
      * @param description  The description for this logical table
      * @param tableGroup  The tablegroup for the logical table
      * @param schema The LogicalTableSchema backing this LogicalTable
+     * @param viewFilters  A list of filters that get attached to any API query sent to this logical table.
      */
     protected LogicalTable(
             @NotNull String name,
@@ -138,7 +160,8 @@ public class LogicalTable implements Table, Comparable<LogicalTable> {
             ReadablePeriod retention,
             String description,
             TableGroup tableGroup,
-            LogicalTableSchema schema
+            LogicalTableSchema schema,
+            ApiFilters viewFilters
     ) {
         this.name = name;
         this.tableGroup = tableGroup;
@@ -148,6 +171,7 @@ public class LogicalTable implements Table, Comparable<LogicalTable> {
         this.description = description;
         this.comparableParam = name + granularity.toString();
         this.schema = schema;
+        this.viewFilters = viewFilters == null ? null : new ApiFilters(viewFilters);
     }
 
     public TableGroup getTableGroup() {
@@ -183,6 +207,10 @@ public class LogicalTable implements Table, Comparable<LogicalTable> {
 
     public Granularity getGranularity() {
         return schema.getGranularity();
+    }
+
+    public Optional<ApiFilters> getFilters() {
+        return Optional.ofNullable(viewFilters);
     }
 
     @Override