+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/TableGroup.java
@@ -4,24 +4,57 @@ package com.yahoo.bard.webservice.table;
 
 import com.yahoo.bard.webservice.data.config.names.ApiMetricName;
 import com.yahoo.bard.webservice.data.dimension.Dimension;
+import com.yahoo.bard.webservice.web.filters.ApiFilters;
 
 import java.util.LinkedHashSet;
 import java.util.Objects;
 import java.util.Set;
 
 /**
- * A TableGroup is a list of schemas.
+ * A TableGroup has a list of Physical Tables and a set of columns representing the collective schema of these tables.
+ * These tables are scanned by query planning to determine a best table.
+ *
+ * The dimensions and apiMetricNames on the TableGroup are only consumed at config time to aid in building
+ * LogicalTable schemas.
+ *
  */
 public class TableGroup {
 
-    private final LinkedHashSet<Dimension> dimensions;
     private final LinkedHashSet<PhysicalTable> tables;
+
+    private final LinkedHashSet<Dimension> dimensions;
     private final Set<ApiMetricName> apiMetricNames;
+
+    private final ApiFilters apiFilters;
 
     /**
      * Builds a TableGroup.
      * A TableGroup contains the dimensions, metrics, and backing physical tables intended to be attached to a
      * LogicalTable.
+     *
+     * @param tables  The backing physical tables
+     * @param apiMetricNames  The metric names for a LogicalTable
+     * @param dimensions  The dimensions for a LogicalTable
+     * @param apiFilters  A collection of ApiFilters to apply to the rows on this table group
+     */
+    public TableGroup(
+            LinkedHashSet<PhysicalTable> tables,
+            Set<ApiMetricName> apiMetricNames,
+            Set<Dimension> dimensions,
+            ApiFilters apiFilters
+    ) {
+        this.tables = tables;
+        this.apiMetricNames = apiMetricNames;
+        this.dimensions = new LinkedHashSet<>(dimensions);
+        this.apiFilters = apiFilters;
+    }
+
+    /**
+     * Builds a TableGroup.
+     * A TableGroup contains the dimensions, metrics, and backing physical tables intended to be attached to a
+     * LogicalTable.
+     *
+     * Default ApiFilters to no filters (null).
      *
      * @param tables  The backing physical tables
      * @param apiMetricNames  The metric names for a LogicalTable
@@ -32,17 +65,16 @@ public class TableGroup {
             Set<ApiMetricName> apiMetricNames,
             Set<Dimension> dimensions
     ) {
-        this.tables = tables;
-        this.apiMetricNames = apiMetricNames;
-        this.dimensions = new LinkedHashSet<>(dimensions);
+        this(tables, apiMetricNames, dimensions, null);
     }
+
 
     /**
      * Getter for set of physical tables.
      *
      * @return physicalTableSchema
      */
-    public Set<PhysicalTable> getPhysicalTables() {
+    public LinkedHashSet<PhysicalTable> getPhysicalTables() {
         return this.tables;
     }
 
@@ -53,6 +85,15 @@ public class TableGroup {
      */
     public Set<Dimension> getDimensions() {
         return dimensions;
+    }
+
+    /**
+     * Getter the row filtering ApiFilters for this table.
+     *
+     * @return apiFilters, null if no filters
+     */
+    public ApiFilters getApiFilters() {
+        return apiFilters;
     }
 
     @Override