+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/data/DruidQueryBuilderSpec.groovy
@@ -11,8 +11,6 @@ import static org.joda.time.DateTimeZone.UTC
 
 import com.yahoo.bard.webservice.data.config.names.DataSourceName
 import com.yahoo.bard.webservice.data.dimension.Dimension
-import com.yahoo.bard.webservice.druid.model.builders.DruidFilterBuilder
-import com.yahoo.bard.webservice.druid.model.builders.DruidOrFilterBuilder
 import com.yahoo.bard.webservice.data.metric.LogicalMetric
 import com.yahoo.bard.webservice.data.metric.LogicalMetricInfo
 import com.yahoo.bard.webservice.data.metric.TemplateDruidQuery
@@ -22,8 +20,13 @@ import com.yahoo.bard.webservice.data.time.ZonedTimeGrain
 import com.yahoo.bard.webservice.data.volatility.DefaultingVolatileIntervalsService
 import com.yahoo.bard.webservice.druid.model.DefaultQueryType
 import com.yahoo.bard.webservice.druid.model.builders.DefaultDruidHavingBuilder
+import com.yahoo.bard.webservice.druid.model.builders.DruidFilterBuilder
+import com.yahoo.bard.webservice.druid.model.builders.DruidOrFilterBuilder
 import com.yahoo.bard.webservice.druid.model.datasource.DefaultDataSourceType
+import com.yahoo.bard.webservice.druid.model.filter.AndFilter
 import com.yahoo.bard.webservice.druid.model.filter.Filter
+import com.yahoo.bard.webservice.druid.model.filter.InFilter
+import com.yahoo.bard.webservice.druid.model.filter.NotFilter
 import com.yahoo.bard.webservice.druid.model.having.Having
 import com.yahoo.bard.webservice.druid.model.orderby.LimitSpec
 import com.yahoo.bard.webservice.druid.model.orderby.OrderByColumn
@@ -35,6 +38,9 @@ import com.yahoo.bard.webservice.druid.model.query.TimeSeriesQuery
 import com.yahoo.bard.webservice.druid.model.query.TopNQuery
 import com.yahoo.bard.webservice.metadata.DataSourceMetadataService
 import com.yahoo.bard.webservice.table.ConstrainedTable
+import com.yahoo.bard.webservice.table.LogicalTable
+import com.yahoo.bard.webservice.table.TableGroup
+import com.yahoo.bard.webservice.table.TableIdentifier
 import com.yahoo.bard.webservice.table.TableTestUtils
 import com.yahoo.bard.webservice.table.resolver.DefaultPhysicalTableResolver
 import com.yahoo.bard.webservice.web.ApiFilter
@@ -77,7 +83,7 @@ public class DruidQueryBuilderSpec extends Specification {
     static final DruidFilterBuilder DRUID_FILTER_BUILDER = new DruidOrFilterBuilder()
 
     List<Interval> intervals
-    static FilterBinders filterBinders = FilterBinders.INSTANCE
+    static FilterBinders filterBinders = FilterBinders.instance
 
 
     def staticInitialize() {
@@ -130,9 +136,12 @@ public class DruidQueryBuilderSpec extends Specification {
         apiRequest.getGranularity() >> HOUR.buildZonedTimeGrain(UTC)
         apiRequest.getTimeZone() >> UTC
         apiRequest.getDimensions() >> ([resources.d1] as Set)
+
+        // Is this correct? All filters use the same dimension, and since we are not merging with the existing set the result only picks up the last filter.
         ApiFilters apiFilters = new ApiFilters(
                 apiFiltersByName.collectEntries {[(resources.d3): [it.value] as Set]} as Map<Dimension, Set<ApiFilter>>
         )
+
         apiRequest.getApiFilters() >> apiFilters
         apiRequest.getLogicalMetrics() >> ([lm1] as Set)
         apiRequest.getIntervals() >> intervals
@@ -528,5 +537,78 @@ public class DruidQueryBuilderSpec extends Specification {
         DefaultQueryType.GROUP_BY   | [:]                               | 1     | false  | 0      | "groupBy"
         DefaultQueryType.GROUP_BY   | [:]                               | 0     | true   | 0      | "groupBy"
         DefaultQueryType.GROUP_BY   | [:]                               | 0     | false  | 1      | "groupBy"
+    }
+
+    def "LogicalTable filters and ApiRequest filters merge properly"() {
+        setup:
+        // inject druid filter builder to test interactions against
+        DruidFilterBuilder dfb = Mock()
+        DruidQueryBuilder builder = new DruidQueryBuilder(
+                resources.logicalDictionary,
+                resolver,
+                dfb,
+                resources.druidHavingBuilder
+        )
+
+        // create logical table with test filter
+        ApiFilter tableFilter = filterBinders.generateApiFilter("ageBracket|id-eq[1,2,3,4]", resources.dimensionDictionary)
+        ApiFilters tableFilters = new ApiFilters([(resources.d3) : [tableFilter] as Set] as Map)
+        LogicalTable baseTable = resources.lt12
+        TableGroup tableGroup = baseTable.getTableGroup();
+        tableGroup = new TableGroup(
+                tableGroup.getPhysicalTables(),
+                tableGroup.getApiMetricNames(),
+                tableGroup.getDimensions(),
+                tableFilters
+        );
+        LogicalTable table = new LogicalTable(
+                baseTable.getName(),
+                baseTable.getCategory(),
+                baseTable.getLongName(),
+                baseTable.getGranularity(),
+                baseTable.getRetention(),
+                baseTable.getDescription(),
+                tableGroup,
+                resources.metricDictionary
+        )
+
+        // create and prep api request
+        apiRequest = Mock(DataApiRequest)
+        apiRequest.getTable() >> table
+        initDefault(apiRequest)
+
+        // put table into table dictionary
+        resources.logicalDictionary.put(TableIdentifier.create(apiRequest), table)
+
+        // prep expected api filters
+        ApiFilter requestFilter = apiRequest.getApiFilters().get(resources.d3).iterator().next()
+        ApiFilters expectedApiFilters = new ApiFilters(
+                [
+                    (resources.d3): [tableFilter, requestFilter] as Set
+                ] as Map
+        )
+
+        when:
+        DruidAggregationQuery daq = builder.buildQuery(apiRequest, resources.simpleTemplateQuery)
+        AndFilter andFilter = (AndFilter) daq.getFilter()
+
+        then: "The combined ApiFilters are used to build the druid filter"
+        1 * dfb.buildFilters({(ApiFilters) it == expectedApiFilters}) >> {
+            ApiFilters filterMap ->
+                System.println(filterMap)
+                resources.druidFilterBuilder.buildFilters(filterMap)
+        }
+
+        and: "it now has the negative assertion from the request"
+
+        andFilter != null
+        andFilter.getFields().any() {
+            it instanceof NotFilter && ((InFilter) ((NotFilter) it).field).values.contains("3")
+        }
+
+        and: "it also contains the positive assertion from the table"
+        andFilter.getFields().any() {
+            it instanceof InFilter && ((InFilter) it).values.equals(["1", "2", "3", "4"] as TreeSet)
+        }
     }
 }